Added INLINE compiler flag
To make sure that Delta method is inlined when possible and to compare against performance difference when it is not.
Although the method has the MethodImpl(MethodImplOptions.AggressiveInlining) attribute, this attribute does not always seem to kick in.
Also commented out use of VectorizedIndexOf.UnsafeIndexOf2 that does not seem to pay off.