Conditional Compilation to SymbolicRegex
Added support for testing if a regex can be compiled to a symbolic regex.
Added also a method TryCompile that tries to compile a regex to a symbolic regex and if this fails, provides a failure reason in an out parameter. Both methods are provided as extension methods of .NET Regex class.
There are two main filure reasons: not implemented constructs (\b \B), and not supported constructs (will not be implemented): \back refereces.