Improved Serialization Support
Fixed serialization bug related to IgnoreCase.
Fixed also serialization related to nonascii that was incorrectly
seralialized when the strings involved misplaced surrogates,
e.g. when a regex is about searching for a single surrogate then
this surrogate has no meaning as a Unicode character.
To avoid this, all serialization of strings inside the matcher serializer
uses now their character code values.

Added RegexMatcher abstract class that implements IMatcher.
This class has Serialize and Deserialize methods and hides the fact
of which actual class is being serialized. Currently it can either be SymbolicRegexUInt64 or SymbolicRegexBV