SymbolicRegex Prettyprinting bugfixes
Fixed bugs in ToString of SymbolicRegex. For eaxmple (a+)? was printed as a+? making it into a lazy loop that has a different semantics than the intended (a+|()).
Added a testsuite simpl.txt for testing internal consistence of the matcher, for this test suite if a match is generated then earliest match should coincides with leftmost match.
Added a separate test that highlights diference bwteeen leftmost and earliest eager matches -- earliest match has the earliest possible end while this endpoint is eagerly extended, e.g. using regex ax+a|x+, the earliest eager match in "zzzaxxxazzz" is xxxx while the leftmost eager match is axxxxa
Exposed some methods as internal methods in the RegexMatcher base class.