Context Free Grammar Intersect with Regex Related
Added better parsing suppoprt for empty rhs, such as
S -> \( S \) | ()
where () is parsed as epsilon

Changed the method name Intersect to IntersectsWith, returning a boolean
and outputting a witness as an output parameter, to more accurately reflect the intent of the method (not to be confused with computing the intersection CFG itself that is currently not supported).

Added more unit tests to cover some more corner cases.