PDA bounded exploration support
Added method Explore(int stackDepth) to PDA that converts the PDA into an automaton that underapproximates the language of the PDA. The PDA is explored so that stacks up to bounded depth are only considered (i.e. up to the given depth bound, unless the depth bound is 0 in which case the exploration may not terminate).